+++ CHANGELOG.md
@@ -27,6 +27,8 @@
   - New functin `OwcJsonService.createLayerFromDefaultOffering()`
 
 ### Bug Fixes
+* **@dlr-eoc/services-ogc:**
+  - `OwcJsonService.createVectorLayerFromOffering` did not work properly with WFS'es: only a substring of the `GetFeature`-operation.href had been used. Now fixed.
 * **@dlr-eoc/map-ol:**
   - Fixed an issue in `MapOlService.zoomInOut()`
   - Fixed an issue with z-index of UkisCustomLayers containing an OlLayerGroup. Before the fix the layergroup's layers would always appear at the very bottom of the map, even below base-layers.